# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Wanted Market API",
    "description": "Wanted Market REST API 명세서\n\n## 인증\n- 대부분의 API는 JWT 토큰 인증이 필요합니다.\n- Authorization 헤더에 'Bearer {token}' 형식으로 토큰을 전달해야 합니다.\n\n## 에러 응답\n- 400: 잘못된 요청\n- 401: 인증 실패\n- 403: 권한 없음\n- 404: 리소스를 찾을 수 없음\n- 409: 충돌 (중복된 데이터 등)\n- 500: 서버 에러",
    "contact": {
      "name": "API Support",
      "email": "support@wantedmarket.com"
    },
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "JWT-Authorization": []
    }
  ],
...
  "paths": {
    "/api/transactions/{id}/status": {
      "put": {
        "tags": [
          "Transaction"
        ],
        "summary": "거래 상태 변경",
        "description": "거래의 상태를 변경합니다. (승인/확정)",
        "operationId": "updateTransactionStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "거래 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json;charset=UTF-8": {
              "schema": {
                "$ref": "#/components/schemas/TransactionStatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "거래를 찾을 수 없음",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoTransactionResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoTransactionResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않은 사용자",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoTransactionResponse"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoTransactionResponse"
                }
              }
            }
          },
          "200": {
            "description": "상태 변경 성공",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-Authorization": []
          }
        ]
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "상품 상세 조회",
        "description": "상품의 상세 정보를 조회합니다. (비회원 가능)",
        "operationId": "getProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "상품 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "404": {
            "description": "상품을 찾을 수 없음",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoProductResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Product"
        ],
        "summary": "상품 수정",
        "description": "상품 정보를 수정합니다. (판매자만 가능)",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "상품 ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json;charset=UTF-8": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "권한 없음 (판매자가 아님)",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoProductResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않은 사용자",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoProductResponse"
                }
              }
            }
          },
          "200": {
            "description": "수정 성공",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "404": {
            "description": "상품을 찾을 수 없음",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoProductResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-Authorization": []
          }
        ]
      }
    },
...
    "/api/transactions": {
      "post": {
        "tags": [
          "Transaction"
        ],
        "summary": "거래 생성",
        "description": "상품 구매를 위한 거래를 생성합니다.",
        "operationId": "createTransaction",
        "requestBody": {
          "content": {
            "application/json;charset=UTF-8": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoTransactionResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않은 사용자",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoTransactionResponse"
                }
              }
            }
          },
          "404": {
            "description": "상품을 찾을 수 없음",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoTransactionResponse"
                }
              }
            }
          },
          "201": {
            "description": "거래 생성 성공",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-Authorization": []
          }
        ]
      }
    },
    "/api/products": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "상품 목록 조회",
        "description": "판매 가능한 상품 목록을 조회합니다. (비회원 가능)",
        "operationId": "getProducts",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "description": "페이지네이션 정보",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "상품 등록",
        "description": "새로운 상품을 등록합니다. (회원만 가능)",
        "operationId": "createProduct",
        "requestBody": {
          "content": {
            "application/json;charset=UTF-8": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoProductResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않은 사용자",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoProductResponse"
                }
              }
            }
          },
          "201": {
            "description": "등록 성공",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-Authorization": []
          }
        ]
      }
    },
...
    "/api/transactions/my/purchases": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "summary": "구매 내역 조회",
        "description": "내가 구매한 거래 목록을 조회합니다.",
        "operationId": "getPurchasedTransactions",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "description": "페이지네이션 정보",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않은 사용자",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoPageTransactionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-Authorization": []
          }
        ]
      }
    },
    "/api/transactions/my/ongoing": {
      "get": {
        "tags": [
          "Transaction"
        ],
        "summary": "진행중인 거래 조회",
        "description": "내가 참여중인 진행중인 거래 목록을 조회합니다.",
        "operationId": "getOngoingTransactions",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "description": "페이지네이션 정보",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "조회 성공",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            }
          },
          "401": {
            "description": "인증되지 않은 사용자",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoPageTransactionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-Authorization": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "TransactionStatusUpdateRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "REQUESTED",
              "APPROVED",
              "CONFIRMED",
              "COMPLETED"
            ]
          }
        },
        "description": "변경할 거래 상태 정보"
      },
      "TransactionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "productId": {
            "type": "integer",
            "format": "int64"
          },
          "productName": {
            "type": "string"
          },
          "buyerId": {
            "type": "integer",
            "format": "int64"
          },
          "buyerName": {
            "type": "string"
          },
          "sellerId": {
            "type": "integer",
            "format": "int64"
          },
          "sellerName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "REQUESTED",
              "APPROVED",
              "CONFIRMED",
              "COMPLETED"
            ]
          },
          "purchasePrice": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResponseDtoTransactionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/TransactionResponse"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ProductUpdateRequest": {
        "required": [
          "price",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "price": {
            "type": "number"
          },
          "quantity": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "수정할 상품 정보"
      },
      "ProductResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "ON_SALE",
              "RESERVED",
              "COMPLETED"
            ]
          },
          "sellerName": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ResponseDtoProductResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/ProductResponse"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "TransactionCreateRequest": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "description": "거래 생성 정보"
      },
      "ProductCreateRequest": {
        "required": [
          "name",
          "price",
          "quantity"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "quantity": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "상품 정보"
      },
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Page": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean"
          },
          "paged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          }
        }
      },
      "PageTransactionResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionResponse"
            }
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "ResponseDtoPageTransactionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/PageTransactionResponse"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "JWT-Authorization": {
        "type": "http",
        "description": "JWT 토큰을 입력해주세요. (Bearer 키워드는 자동으로 추가됩니다.)",
        "name": "JWT Authorization",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}