# src/main/java/com/wanted/market/domain/product/dto/ProductResponse.java
package com.wanted.market.domain.product.dto;

import com.wanted.market.domain.product.Product;
import com.wanted.market.domain.product.ProductStatus;
import lombok.Builder;
import lombok.Getter;

import java.math.BigDecimal;
import java.time.LocalDateTime;

@Getter
@Builder
public class ProductResponse {
    private Long id;
    private String name;
    private BigDecimal price;
    private ProductStatus status;
    private String sellerName;
    private Integer quantity;
    private LocalDateTime createdAt;
    private LocalDateTime updatedAt;

    public static ProductResponse from(Product product) {
        return ProductResponse.builder()
                .id(product.getId())
                .name(product.getName())
                .price(product.getPrice())
                .status(product.getStatus())
                .sellerName(product.getSeller().getName())
                .quantity(product.getQuantity())
                .createdAt(product.getCreatedAt())
                .updatedAt(product.getUpdatedAt())
                .build();
    }
}